+++ advanced-order-platform/src/test/java/com/codexdemo/orderplatform/ops/maintenance/ciarc/OpsShardReadinessMinimalReadOnlyGateOperatorCiHandoffArchiveDigestConsumerPackageVerificationDossierReleaseAcceptanceArchiveRegistryCiBoundaryTests.java
@@ -8,9 +8,7 @@ class OpsShardReadinessMinimalReadOnlyGateOperatorCiHandoffArchiveDigestConsumer
 
   @Test
   void preservesReadOnlyCiAttestations() {
-    var response =
-        OpsShardReadinessMinimalReadOnlyGateOperatorCiHandoffArchiveDigestConsumerPackageVerificationDossierReleaseAcceptanceArchiveRegistryTestSupport
-            .registry();
+    var response = ArchiveTestData.registry();
 
     assertThat(response.ciAttestationCount()).isEqualTo(5);
     assertThat(response.passedCiAttestationCount()).isEqualTo(5);
@@ -31,9 +29,7 @@ class OpsShardReadinessMinimalReadOnlyGateOperatorCiHandoffArchiveDigestConsumer
 
   @Test
   void locksBoundarySealsWithoutOpeningRuntimeExecution() {
-    var response =
-        OpsShardReadinessMinimalReadOnlyGateOperatorCiHandoffArchiveDigestConsumerPackageVerificationDossierReleaseAcceptanceArchiveRegistryTestSupport
-            .registry();
+    var response = ArchiveTestData.registry();
 
     assertThat(response.boundarySealCount()).isEqualTo(8);
     assertThat(response.lockedBoundarySealCount()).isEqualTo(8);