+++ advanced-order-platform/src/test/java/com/codexdemo/orderplatform/ops/maintenance/ciarc/OpsShardReadinessMinimalReadOnlyGateOperatorCiHandoffArchiveDigestConsumerPackageVerificationDossierReleaseAcceptanceArchiveRegistryRouteOperatorTests.java
@@ -8,9 +8,7 @@ class OpsShardReadinessMinimalReadOnlyGateOperatorCiHandoffArchiveDigestConsumer
 
   @Test
   void carriesRoutePackagesFromReleaseAcceptanceSignoffLanes() {
-    var response =
-        OpsShardReadinessMinimalReadOnlyGateOperatorCiHandoffArchiveDigestConsumerPackageVerificationDossierReleaseAcceptanceArchiveRegistryTestSupport
-            .registry();
+    var response = ArchiveTestData.registry();
 
     assertThat(response.routePackageCount()).isEqualTo(4);
     assertThat(response.readyRoutePackageCount()).isEqualTo(4);
@@ -34,9 +32,7 @@ class OpsShardReadinessMinimalReadOnlyGateOperatorCiHandoffArchiveDigestConsumer
 
   @Test
   void buildsOperatorPacksWithStableOrdering() {
-    var response =
-        OpsShardReadinessMinimalReadOnlyGateOperatorCiHandoffArchiveDigestConsumerPackageVerificationDossierReleaseAcceptanceArchiveRegistryTestSupport
-            .registry();
+    var response = ArchiveTestData.registry();
 
     assertThat(response.operatorPackCount()).isEqualTo(4);
     assertThat(response.readyOperatorPackCount()).isEqualTo(4);